[feature/#6] 로그아웃 구현
