+++ src/main/java/com/api/jaebichuri/auth/controller/AuthController.java
@@ -10,7 +10,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -47,11 +46,5 @@ public class AuthController {
         Map<String, String> memberInfo = authService.getMemberInfo(code);
         return ResponseEntity.ok(
             ApiResponse.onSuccess(memberService.successSocialLogin(memberInfo)));
-    }
-
-    @GetMapping("/reissue")
-    ResponseEntity<ApiResponse<TokenResponseDto>> reissue(
-        @RequestAttribute("refresh") String refreshToken) {
-        return ResponseEntity.ok(ApiResponse.onSuccess(authService.reissue(refreshToken)));
     }
 }
